bpf: refactor parser and enable XDP.
The patch does the following:
1) Refactor the original parser into data parsing
   and metadata parsing
2) Move the packet data parsing code to parser_common.h
3) Enable xdp.h to parse packet

As a result, TC and XDP share the same packet data parsing
code.  If a netdev supports XDP, the packet will be first
parsed at XDP, save in percpu_headers map, and passes to the
upper TC metadata parsing (percpu_metadata map). If not, then
the TC does full parsing.
